Update admin keywords and node-sass
This is how they have been harvested, in python notation:

  crate_admin_keywords = set([
     copy&paste keywords crate-admin/app/scripts/controllers/console.js
  ])
  crate_symbolic_keywords = set(map(lambda t: t.lower(), (
     copy&paste keywords _SYMBOLIC_NAMES crate SqlBaseLexer.java
  )))
  new_terms = crate_symbolic_keywords.difference(crate_admin_keywords)
  final_terms = crate_admin_keywords.union(new_terms)

  def show(title, col, max_chars_in_line=60):
    sorted_terms = list(col)
    sorted_terms.sort()
    print("{}:".format(title))
    line_offset = 1
    for t in sorted_terms:
        print("'{}', ".format(t), end='')
        line_offset += 4 + len(t)
        if line_offset > max_chars_in_line:
            line_offset = 1
            print("")
    print("\n{}".format("=" * 100))
  show(final_terms)

Update node-sass (links to a C library for spreadsheets) to 4.13.0,
a minor release. This was done automatically when I built/run the
application. The release only contains typo corrections and other
minor config related things, which do not affect the behaviour of
the AdminUI (https://github.com/sass/node-sass/releases/tag/v4.13.0).