Remove misguided redirects which may lose queries currently being typed
Users noticed a UI glitch when using the UI console. When working on a
query in the console view, the interface frequently jumps back to
displaying the first table of the first schema in the "Tables view".
This behaviour has been observed when working on queries related to
subordinate schemas.

This issue seems to be unrelated to the query itself. It occurs before
and also while editing the query in the UI console, but not after the
query was executed. This is especially inconvenient when longer queries
get lost.